move workflow support into build package
