+++ lkbuild/__init__.py
@@ -1,0 +1,4 @@
+"""
+LensKit build bootstrap utilities.  This repository should generally not be
+installed.
+"""
